requirements.txt
altair==5.3.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.1.0
asttokens==2.2.1
attrs==21.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
beniget==0.4.1
bleach==4.1.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
cachetools==5.3.3
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
comm==0.1.4
command-not-found==0.3
contourpy==1.2.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.6.7.post1
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
dnspython==2.6.1
duplicity==0.8.21
entrypoints==0.4
executing==1.2.0
fasteners==0.14.1
filelock==3.13.0
fonttools==4.45.1
frozendict==2.3.8
fs==2.4.12
fsspec==2023.10.0
future==0.18.2
gast==0.5.2
gitdb==4.0.11
GitPython==3.1.43
google==3.0.0
googlesearch-python==1.2.3
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.25.1
ipython==8.14.0
ipython_genutils==0.2.0
ipywidgets==6.0.0
jedi==0.19.0
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==3.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-pygments==0.1.2
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.3
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multitasking==0.0.11
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.7
netifaces==0.11.0
networkx==3.2.1
notebook==6.4.8
numpy==1.25.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
packaging==23.1
pandas==2.1.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
peewee==3.17.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==3.10.0
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.39
protobuf==4.25.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pyarrow==16.0.0
pycairo==1.20.1
pycups==2.0.1
pydeck==0.8.1b0
Pygments==2.16.1
PyGObject==3.42.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.3.0+ubuntu2
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-whois==0.9.3
pythran==0.10.0
pytz==2023.3.post1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.1
reportlab==3.6.8
requests==2.31.0
rich==13.7.1
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.0
SecretStorage==3.3.1
Send2Trash==1.8.1b0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
ssh-import-id==5.11
stack-data==0.6.2
streamlit==1.33.0
sympy==1.12
systemd-python==234
tenacity==8.2.3
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.1
torch==2.1.0
tornado==6.3.3
traitlets==5.9.0
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
watchdog==4.0.0
wcwidth==0.2.6
webencodings==0.5.1
whois==1.20240129.2
widgetsnbextension==2.0.0
xdg==5
xkit==0.0.0
yfinance==0.2.31
zipp==1.0.0
